+++ src/main/java/com/kodilla/ecommercee/controller/GroupController.java
@@ -1,0 +1,5 @@
+package com.kodilla.ecommercee.controller;
+
+public class GroupController {
+
+}
